syz-fuzzer: remove testing mode
It's not used anymore.
